+++ python/ray/tests/test_runtime_env.py
@@ -13,7 +13,6 @@ import ray.experimental.internal_kv as kv
 from ray._private.test_utils import (
     run_string_as_driver, run_string_as_driver_nonblocking, wait_for_condition)
 from ray._private.runtime_env import working_dir as working_dir_pkg
-from ray._private.runtime_env.validation import override_task_or_actor_runtime_env  # noqa: E501
 from ray._private.utils import (get_wheel_filename, get_master_wheel_url,
                                 get_release_wheel_url)
 
@@ -941,39 +940,6 @@ def test_large_file_error(shutdown_only):
             ray.init(runtime_env={"working_dir": "."})
 
         os.chdir(old_dir)
-
-
-class TestOverrideTaskOrActorRuntimeEnv:
-    def test_uri_inherit(self):
-        child_env = {}
-        parent_env = {"working_dir": "other_dir", "uris": ["a", "b"]}
-        result_env = override_task_or_actor_runtime_env(child_env, parent_env)
-        assert result_env == {"uris": ["a", "b"]}
-
-        # The dicts passed in should not be mutated.
-        assert child_env == {}
-        assert parent_env == {"working_dir": "other_dir", "uris": ["a", "b"]}
-
-    def test_uri_override(self):
-        child_env = {"uris": ["c", "d"]}
-        parent_env = {"working_dir": "other_dir", "uris": ["a", "b"]}
-        result_env = override_task_or_actor_runtime_env(child_env, parent_env)
-        assert result_env["uris"] == ["c", "d"]
-        assert result_env.get("working_dir") is None
-
-        # The dicts passed in should not be mutated.
-        assert child_env == {"uris": ["c", "d"]}
-        assert parent_env == {"working_dir": "other_dir", "uris": ["a", "b"]}
-
-    def test_no_mutate(self):
-        child_env = {}
-        parent_env = {"working_dir": "other_dir", "uris": ["a", "b"]}
-        result_env = override_task_or_actor_runtime_env(child_env, parent_env)
-        assert result_env == {"uris": ["a", "b"]}
-
-        # The dictis passed in should not be mutated.
-        assert child_env == {}
-        assert parent_env == {"working_dir": "other_dir", "uris": ["a", "b"]}
 
 
 if __name__ == "__main__":
